turned game container and associated logic into a separate component
